requirements.txt
pyzotero
openai
requests
beautifulsoup4
pyyaml
arxiv
tqdm